Added infrastructure for multiple recipe
