+++ package.json
@@ -25,6 +25,7 @@
         "near-api-js": "^0.44.2",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
+        "react-router-dom": "^6.2.2",
         "react-scripts": "5.0.0",
         "styled-components": "^5.3.3",
         "typescript": "^4.4.2",
