+++ src/App.tsx
@@ -1,15 +1,17 @@
 import CssBaseline from "@mui/material/CssBaseline"
 import { createTheme, ThemeProvider } from "@mui/material/styles"
 import TimelineComponent from "./components/timeline"
-import PaperComponent from "./components/paper"
 import SummaryBox from "./components/summaryBox"
 import Header from "./components/header"
-import { Box, Grid } from "@mui/material"
+import { Box, Button, Grid, Icon, Paper } from "@mui/material"
 import { useReducer } from "react"
 import { Refresh } from "./utils/refresh"
 import WalletComponent from "./components/wallet"
 import { initNear } from "./services/near"
 import { getPage } from "./utils/navigation"
+import { NavLink, Outlet, useParams } from "react-router-dom"
+import PageComponent from "./components/page"
+import { recipes } from "./recipes/recipes"
 
 declare module "@mui/material/styles/createPalette" {
     interface Palette {
@@ -24,8 +26,6 @@ declare global {
     interface Window {
         updateApp: any
         updatePage: any
-        REFRESHER: Refresh[]
-        EMPTY_REFRESH: Refresh
     }
 }
 
@@ -66,9 +66,6 @@ const theme = createTheme({
 window.nearInitPromise = initNear().then(window.updateApp)
 
 export default function App() {
-    const [, forceUpdate] = useReducer(x => x + 1, 0)
-    window.updateApp = forceUpdate
-    const page = getPage()
     return (
         <ThemeProvider theme={theme}>
             <Box
@@ -81,51 +78,122 @@ export default function App() {
             >
                 <CssBaseline />
                 <Header />
-                {page === 0 ? <SummaryBox /> : <></>}
-                <Grid
-                    container
-                    sx={{
-                        width: 1,
-                        height: 1,
-                        flex: "1 1 0"
-                    }}
-                    direction="row"
-                    justifyContent="center"
-                    alignItems="center"
-                    wrap="nowrap"
-                    position="sticky"
-                >
-                    <Grid
-                        item
-                        container
-                        direction="column"
-                        alignItems="flex-end"
-                        wrap="nowrap"
-                        flexShrink={0}
-                        sx={{
-                            height: "fit-content"
-                        }}
-                        xs={2}
-                    >
-                        <WalletComponent />
-                        <TimelineComponent />
-                    </Grid>
-                    <Grid
-                        item
-                        sx={{
-                            height: 0.5,
-                            minHeight: "500px",
-                            maxHeight: "800px",
-                            flexBasis: "800px",
-                            flexShrink: 0,
-                            width: 0.5
-                        }}
-                    >
-                        <PaperComponent />
-                    </Grid>
-                    <Grid item xs={2} sx={{ height: 0, flex: "1 1 0 !important" }} />
-                </Grid>
+                <Outlet />
             </Box>
         </ThemeProvider>
+    )
+}
+
+export function RecipePage() {
+    const params = useParams()
+    const page = getPage()
+    const [, forceUpdate] = useReducer(x => x + 1, 0)
+    window.updateApp = forceUpdate
+    return (
+        <>
+            {page === 0 ? <SummaryBox /> : <></>}
+            <Grid
+                container
+                sx={{
+                    width: 1,
+                    height: 1,
+                    flex: "1 1 0"
+                }}
+                direction="row"
+                justifyContent="center"
+                alignItems="center"
+                wrap="nowrap"
+                position="sticky"
+            >
+                <Grid
+                    item
+                    container
+                    direction="column"
+                    alignItems="flex-end"
+                    wrap="nowrap"
+                    flexShrink={0}
+                    sx={{
+                        height: "fit-content"
+                    }}
+                    xs={2}
+                >
+                    <WalletComponent />
+                    <TimelineComponent />
+                </Grid>
+                <Grid
+                    item
+                    sx={{
+                        height: 0.5,
+                        minHeight: "500px",
+                        maxHeight: "800px",
+                        flexBasis: "800px",
+                        flexShrink: 0,
+                        width: 0.5
+                    }}
+                >
+                    <Paper
+                        sx={{
+                            width: 1,
+                            height: "fit-content",
+                            minHeight: 1,
+                            mb: 4,
+                            display: "flex",
+                            "& > *": {
+                                height: "unset !important"
+                            }
+                        }}
+                        elevation={2}
+                    >
+                        <PageComponent recipe={parseInt(params.recipeId!)} page={getPage()} />
+                    </Paper>
+                </Grid>
+                <Grid item xs={2} sx={{ height: 0, flex: "1 1 0 !important" }} />
+            </Grid>
+        </>
+    )
+}
+
+export function CatalogPage() {
+    return (
+        <Grid
+            container
+            sx={{
+                width: 1,
+                height: 1,
+                flex: "1 1 0",
+                overflowY: "scroll"
+            }}
+            direction="column"
+            alignItems="center"
+            wrap="nowrap"
+        >
+            {recipes.map(r => (
+                <Paper
+                    sx={{
+                        width: 0.5,
+                        maxWidth: "800px",
+                        height: "min-content",
+                        mb: 4,
+                        p: 2,
+                        display: "flex",
+                        flexFlow: "column nowrap",
+                        position: "relative"
+                    }}
+                    elevation={2}
+                >
+                    <h3 style={{marginTop: 0}}>{r.title}</h3>
+                    <div>{r.description}</div>
+                    <NavLink to={`/${r.id}`} key={r.id}>
+                        <Button
+                            variant="outlined"
+                            sx={{ borderRadius: "100px", position: "absolute", right: "16px", bottom: "16px" }}
+                            endIcon={<Icon>navigate_next</Icon>}
+                        >
+                            START
+                        </Button>
+                    </NavLink>
+                </Paper>
+            ))}
+        </Grid>
     )
 }
